requirements.txt
mcdreforged>=2.1
pycryptodome
